+++ src/get-started/get-started.scss
@@ -1,0 +1,8 @@
+.get-started-container{
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  .get-started{
+    max-width: 75ch;
+  }
+}
